refactor: Split TxConfirmLayout into TxLayout and TxSign
