+++ source/pages/Transactions/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { TxConfirmLayout, TxLayout } from 'components/Layout';
+import { TxLayout, TxSign } from 'components/Layout';
 
 //* CreateToken
 export const Create = () => <TxLayout txType="CreateToken" />;
@@ -12,19 +12,10 @@ export const CreateAndIssueNFT = () => <TxLayout txType="CreateNFT" />;
 export const MintNFT = () => <TxLayout txType="MintNFT" />;
 
 //* SignAndSend
-export const SignAndSend = () => (
-  <TxConfirmLayout
-    sign
-    signAndSend
-    txType="SignAndSend"
-    title="Signature Request"
-  />
-);
+export const SignAndSend = () => <TxSign send />;
 
 //* SignPSBT
-export const SignPSBT = () => (
-  <TxConfirmLayout sign txType="Sign" title="Signature Request" />
-);
+export const SignPSBT = () => <TxSign />;
 
 //* TransferOwnership
 export const TransferOwnership = () => <TxLayout txType="TransferToken" />;
